requirements.txt
bcj-cffi==0.5.1
Brotli==1.0.9
cffi==1.14.6
click==8.0.1
colorama==0.4.4
EasyProcess==0.3
entrypoint2==0.2.4
importlib-metadata==4.6.1
multivolumefile==0.2.3
patool==1.12
Pillow==8.3.1
py7zr==0.16.1
pycparser==2.20
pycryptodomex==3.10.1
pyppmd==0.15.2
pyunpack==0.2.2
pyzstd==0.14.4
texttable==1.6.4
typing-extensions==3.10.0.0
zipp==3.5.0
